Backnd/requirements.txt
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.2.post1
asyncpg==0.30.0
bcrypt==4.0.1
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.1
dnspython==2.7.0
ecdsa==0.19.0
email-validator==2.0.0.post2
fastapi==0.115.2
greenlet==3.1.1
h11==0.14.0
idna==3.10
Mako==1.3.6
MarkupSafe==3.0.2
passlib==1.7.4
psycopg2-binary==2.9.10
pyasn1==0.4.8
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pyotp==2.8.0
python-dotenv==1.0.0
python-jose==3.4.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.40.0
typing_extensions==4.12.2
uvicorn==0.34.0